Drop _Impl class name suffix where pointless
I.e. where there is no other LibreOffice, or system, or 3rd-party
library thing with just the un-suffixed name that they would be a
pimpl of, etc. Why bother calling something InterfaceOleWrapper_Impl
when there is nothing that would be called InterfaceOleWrapper? It
just makes the code seem more complicated and harder to read.

An "implementation" of something is what everything that is not fully
abstract *is*. We don't suffix the name of every non-abstract class
with _Impl elsewhere either.

Also drop two unused structs.

Change-Id: I2755e6da78d8e4ded5291688dc90510b5f8b498b
Reviewed-on: https://gerrit.libreoffice.org/54411
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tor Lillqvist <tml@collabora.com>