tdf#117440 Can't move focus to disposed window
xSaveFocusId is the already disposed comments floating window,
so it can't accept focus anymore. Moreover, if closing the
menu by clicking its title, the focus results nowhere, and no
keyboard interaction works at all (including e.g. F6 or F10).

This patch makes the focus switch back to the document, like
it was in earlier LO versions. And does it in a way that
doesn't break tdf#115227.

Change-Id: I1d473557c10faf65b5b1fa917f5a5d66d0de8656
Reviewed-on: https://gerrit.libreoffice.org/55457
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>