NFC sd misc cleanup: earlier exit, comments
The exception in the case of a running slideshow has been around
since the function was introduced, but more and more code
has been building up in front of that if clause. Since no
changes will be made, include that exception in the initial
exit clause to avoid unnecessary computations.

This cleanup is prep work for a bug fix.

Change-Id: Ief0b07d8108e1576459486ef61f038443f64705f
Reviewed-on: https://gerrit.libreoffice.org/54678
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>