disable calc's threading for external references
Since the code uses ScExternalRefManager, which caches values, so
better just disable it. Can be seen e.g. with fdo#78490-1 (and
mnOpenCLMinimumFormulaGroupSize disabled). The doc actually asserts
in GetFormatTable() called from ScExternalRefManager, but there
seems to be no point in fixing just the assert if the whole class
is possibly problematic.

Change-Id: I4ace488414fd15b4ad0b88da51205b02c561c7a6
Reviewed-on: https://gerrit.libreoffice.org/56289
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Luboš Luňák <l.lunak@collabora.com>